팀장 담당 부분, 재배정 예정
