+++ add/add.html
@@ -7,17 +7,16 @@
     <title>Document</title>
 </head>
 <body>
-    <form action="/detail" method="post">
-        <input type="text" name="id" placeholder="키"><p></p>
-        
+    <form action="/add" method="post">
+    
     <!-- 제목 -->
         <input type="text" name="title" placeholder="제목"><p></p>
         
     <!-- 업로드 -->
-        <!-- <input type="file" name="data"><p></p> -->
+        <input type="file" name="data"><p></p>
 
     <!-- 상세 -->
-        <input type="text" name="write" placeholder="내용"><p></p>
+        <input type="text" name="content" placeholder="내용"><p></p>
 
     <!-- 제출 -->
         <input type="submit"><p></p>
