+++ routes/cart.js
@@ -3,7 +3,7 @@ const router = express.Router()
 
 
 router.get('/hi', async(req,res) => {
-    const dbData = await test
+    const dbData = await test()
     res.send("hi")
 })
